gitdub: add ability to set name on env sender
Git's post-receive-mail script uses sendmail to send, which uses the
-f argument to set the envelope sender.  However, this doesn't
recognize the form "foo <bar@example.com>", meaning that this form
will use "<bar@example.com>" and will ignore the "foo" part.  So strip
out the "foo" part (if it's there) and set that in the NAME enviroment
variable, which sendmail will recognize.

Signed-off-by: Jeff Squyres <jeff@squyres.com>